QEDcore

Documentation for QEDcore.

QEDcore.AbstractAxisBeta

QEDcore.AbstractAxisBoostParameter

QEDcore.AbstractBoostParameter

QEDcore.AbstractBoostVector

QEDcore.AbstractCoordinateTransformation

QEDcore.AbstractLorentzBoost

QEDcore.AbstractLorentzTransformation

QEDcore.AdjointBiSpinor

QEDcore.AntiBoson

QEDcore.AntiFermion

QEDcore.BetaVector

QEDcore.BetaX

QEDcore.BetaY

QEDcore.BetaZ

QEDcore.BiSpinor

QEDcore.Boost

QEDcore.Boson

QEDcore.BosonLike

QEDcore.CenterOfMomentumFrame

QEDcore.DiracMatrix

QEDcore.Electron

QEDcore.ElectronRestFrame

QEDcore.Fermion

QEDcore.FermionLike

QEDcore.InPhaseSpacePoint

QEDcore.InPhaseSpacePoint

QEDcore.InPhaseSpacePoint

QEDcore.InPhaseSpacePoint

QEDcore.MFourMomentum

QEDcore.MFourMomentum

QEDcore.MLorentzVector

QEDcore.MajoranaBoson

QEDcore.MajoranaFermion

QEDcore.OutPhaseSpacePoint

QEDcore.OutPhaseSpacePoint

QEDcore.OutPhaseSpacePoint

QEDcore.ParticleStateful

QEDcore.ParticleStateful

QEDcore.PhaseSpacePoint

QEDcore.PhaseSpacePoint

QEDcore.PhaseSpacePoint

QEDcore.PhasespaceDefinition

QEDcore.Photon

QEDcore.Positron

QEDcore.SFourMomentum

QEDcore.SFourMomentum

QEDcore.SLorentzVector

QEDcore.SphericalCoordinateSystem

Base.getindex

Base.getindex

QEDcore._momentum_type

QEDcore._mul

QEDcore._mul

QEDcore._mul

QEDcore._mul

QEDcore._mul

QEDcore._mul

QEDcore._transform

QEDcore.AbstractAxisBeta — Type

AbstractAxisBeta{T} <: AbstractAxisBoostParameter{T}

An abstract base type for the beta (velocity) parameter of type T, representing a Lorentz boost along a specific spatial axis.

AbstractAxisBeta{T} extends AbstractAxisBoostParameter{T} and provides a general framework for defining beta parameters associated with individual Cartesian axes (x, y, z) in special relativity. The parameter T typically represents the numeric type (e.g., Float64, Float32) used for the beta value.

Usage

Concrete subtypes of AbstractAxisBeta{T} define the beta parameters for Lorentz boosts along the x, y, and z axes:

BetaX{T}: Boost parameter for the x-axis.

BetaY{T}: Boost parameter for the y-axis.

BetaZ{T}: Boost parameter for the z-axis.

These beta parameters are essential for performing axis-specific Lorentz boosts, which transform physical quantities such as four-momentum between different inertial frames.

QEDcore.AbstractAxisBoostParameter — Type

AbstractAxisBoostParameter{T}

An abstract base type representing a boost parameter of type T, associated with a specific axis in space (e.g., $x$, $y$, or $z$).

This type serves as a foundation for concrete boost parameter types that define Lorentz boosts along individual spatial directions. The parameter T typically represents the data type for the boost value (e.g., Float64, Float32).

Usage

Subtypes of AbstractAxisBoostParameter{T} are used to define specific boost transformations along a given axis (such as BetaX for the x-axis). These types are essential in performing Lorentz boosts, which transform four-momentum vectors between different inertial reference frames.

This abstract type is meant to be extended by concrete types to represent boosts along different Cartesian axes.

QEDcore.AbstractBoostParameter — Type

AbstractBoostParameter

An abstract base type representing boost parameters used in Lorentz transformations, which describe the relative motion between two inertial frames in special relativity.

AbstractBoostParameter serves as the foundation for defining specific boost parameters that control Lorentz boosts in different spatial directions. Boost parameters typically represent the velocity of one reference frame relative to another, expressed as a fraction of the speed of light (\beta), and are essential for performing Lorentz transformations on four-vectors (such as SFourMomentum).

Overview

In the context of special relativity, a Lorentz boost is a transformation that changes the time and spatial components of a four-vector based on the relative motion between two inertial reference frames. For example, the boost parameter $\beta$ is dimensionless and represents this velocity as a fraction of the speed of light. Depending on the frame's relative velocity, different forms of boost parameters exist, such as those associated with a single axis or a vector describing boosts in multiple spatial dimensions.

The AbstractBoostParameter type is the parent type for all specific kinds of boost parameters, including:

Axis-specific Boost Parameters: Such as BetaX, which describes a boost along the x-axis.

Vector-like Boost Parameters: Such as BetaVector, which describes boosts with components in multiple spatial directions.

QEDcore.AbstractBoostVector — Type

AbstractBoostVector <: AbstractBoostParameter

An abstract base type representing vector-like boost parameters, used to model Lorentz boosts in any spatial dimension.

AbstractBoostVector extends AbstractBoostParameter and provides the framework for describing boosts that act in multiple spatial dimensions simultaneously, typically in three-dimensional space. This type is designed to support vector representations of velocities (in units of the speed of light) associated with Lorentz transformations in special relativity.

Usage

Concrete subtypes of AbstractBoostVector represent specific boost vectors that describe the velocity components in each spatial dimension, such as BetaVector. These boost vectors are commonly used in transformations of four-vectors (e.g., four-momentum, four-position) between different reference frames.

For example:

BetaVector{T}: A concrete subtype representing a boost vector with velocity components $\beta_x$, $\beta_y$, and $\beta_z$ (in units of the speed of light).

QEDcore.AbstractCoordinateTransformation — Type

AbstractCoordinateTransformation

Abstract base type for coordinate transformations supposed to be acting on four-momenta. Every subtype of trafo::AbstractCoordinateTransformation should implement the following interface functions:

QEDcore._transform(trafo,p): transforms p

Base.inv(trafo): returns the inverted transform

Example

Implementing the interface by defining the interface functions:

julia> using QEDcore

julia> struct TestTrafo{T} <: QEDcore.AbstractCoordinateTransformation
           a::T
       end

julia> QEDcore._transform(trafo::TestTrafo,p) = trafo.a*p

julia> Base.inv(trafo::TestTrafo) = TestTrafo(inv(trafo.a))

The TestTrafo can then be used to transform four-momenta:

julia> trafo = TestTrafo(2.0)
TestTrafo{Float64}(2.0)

julia> p = SFourMomentum(4,3,2,1)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> trafo(p) # multiply every component with 2.0
4-element SFourMomentum with indices SOneTo(4):
 8.0
 6.0
 4.0
 2.0

julia> inv(trafo)(p) # divide every component by 2.0
4-element SFourMomentum with indices SOneTo(4):
 2.0
 1.5
 1.0
 0.5

QEDcore.AbstractLorentzBoost — Type

AbstractLorentzBoost <: AbstractLorentzTransformation

An abstract base type representing Lorentz boosts, a specific type of Lorentz transformation associated with relative motion between inertial frames along one or more spatial directions.

AbstractLorentzBoost extends AbstractLorentzTransformation and serves as the foundation for all types of boost transformations in special relativity. Lorentz boosts describe how four-vectors (such as SFourMomentum) change when transitioning between two reference frames moving at constant velocities (in units of the speed of light) relative to each other.

For example:

Boost{T}: A concrete implementation of Lorentz boosts with boost parameter T (see also AbstractBoostParameter).

QEDcore.AbstractLorentzTransformation — Type

AbstractLorentzTransformation <: AbstractCoordinateTransformation

An abstract base type representing Lorentz transformations, which are coordinate transformations between inertial and reference frames in special relativity.

AbstractLorentzTransformation extends AbstractCoordinateTransformation and provides the foundational framework for all types of Lorentz transformations, including boosts. These transformations preserve the Minkowski product of two four-vectors and are fundamental to the description of relativistic physics, ensuring the laws of physics are the same in all inertial frames.

Usage

Subtypes of AbstractLorentzTransformation implement specific kinds of Lorentz transformations. For example:

Boost{T}: A concrete implementation of Lorentz boosts with boost parameter T (see also AbstractBoostParameter).

These subtypes perform transformations on four-vectors (such as SFourMomentum) between different inertial reference frames.

QEDcore.AdjointBiSpinor — Type

struct AdjointBiSpinor <: AbstractDiracVector{ComplexF64}

Concrete type to model an adjoint Dirac four-spinor with complex-valued components. These are the elements of the dual spinor space.

QEDcore.AntiBoson — Type

Abstract base type for anti-bosons as distinct from its particle counterpart Boson.

particle interface

All subtypes of AntiBoson have

is_boson(::AntiBoson) = true
is_particle(::AntiBoson) = false
is_anti_particle(::AntiBoson) = true

QEDcore.AntiFermion — Type

Abstract base type for anti-fermions as distinct from its particle counterpart Fermion.

particle interface

All subtypes of AntiFermion have

is_fermion(::AntiFermion) = true
is_particle(::AntiFermion) = false
is_anti_particle(::AntiFermion) = true

QEDcore.BetaVector — Type

BetaVector(x::Real,y::Real,z::Real)

Represents the spatial vector of velocity parameters (denoted as the "beta" vector) associated with motion in the three Cartesian directions, i.e., $\vec\beta = (\beta_x, \beta_y, \beta_z)$. These components correspond to the velocity of an object (in units of the speed of light) in each of the $x$, $y$, and $z$ directions.

The Lorentz boost along the direction of the beta vector $\vec\beta$ transforms the four-momentum as follows:

\[\begin{pmatrix} p_0\\ p_1\\ p_2\\ p_3 \end{pmatrix} \mapsto \begin{pmatrix} \gamma (p_0 - \vec\beta \vec p)\\ p_1 + (\frac{\gamma - 1}{\beta^2} \vec\beta\vec p - \gamma p_0) \beta_x\\ p_2 + (\frac{\gamma - 1}{\beta^2} \vec\beta\vec p - \gamma p_0) \beta_y\\ p_3 + (\frac{\gamma - 1}{\beta^2} \vec\beta\vec p - \gamma p_0) \beta_z\\ \end{pmatrix}\]

where the kinematic factor is given as $\gamma = 1/\sqrt{1-\beta_x^2}$.

Example

julia> using QEDcore

julia> beta_vec = BetaVector(0.2,0.3,0.1)
BetaVector{Float64}(0.2, 0.3, 0.1)

julia> boost = Boost(beta_vec)
Boost{BetaVector{Float64}}(BetaVector{Float64}(0.2, 0.3, 0.1))

julia> p = SFourMomentum(4.0,3.0,2.0,1.0)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> p_prime = boost(p)
4-element SFourMomentum with indices SOneTo(4):
 2.911484876492837
 2.282803602436349
 0.9242054036545237
 0.6414018012181746

julia> @assert isapprox(p*p,p_prime*p_prime) # Invariant mass is preserved

External link

Lorentz Boost on Wikipedia

Kinematics in PDG review

ROOT::Math:Boost from ROOT

QEDcore.BetaX — Type

BetaX(beta::T) where {T<:Real}

Represents the beta parameter associated with a Lorentz boost along the x-axis, commonly denoted as $\beta_x$.

The transformation for a boost along the x-axis is:

\[\begin{pmatrix} p_0\\ p_1\\ p_2\\ p_3 \end{pmatrix} \mapsto \begin{pmatrix} \gamma (p_0 - \beta_x p_1)\\ \gamma (p_1 - \beta_x p_0)\\ p_2\\ p_3 \end{pmatrix}\]

where the kinematic factor is given as $\gamma = 1/\sqrt{1-\beta_x^2}$)

Example

julia> using QEDcore

julia> beta_x = BetaX(0.5)
BetaX{Float64}(0.5)

julia> boost = Boost(beta_x)
Boost{BetaX{Float64}}(BetaX{Float64}(0.5))

julia> p = SFourMomentum(4,3,2,1)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> p_prime = boost(p)
4-element SFourMomentum with indices SOneTo(4):
 2.886751345948129
 1.1547005383792517
 2.0
 1.0

julia> @assert isapprox(p*p,p_prime*p_prime) # Invariant mass is preserved

External link

Lorentz Boost on Wikipedia

Kinematics in PDG review

QEDcore.BetaY — Type

BetaY(beta::T) where {T<:Real}

Represents the beta parameter associated with a Lorentz boost along the y-axis, commonly denoted as $\beta_y$.

The transformation for a boost along the y-axis is:

\[\begin{pmatrix} p_0\\ p_1\\ p_2\\ p_3 \end{pmatrix} \mapsto \begin{pmatrix} \gamma (p_0 - \beta_y p_2)\\ p_1\\ \gamma (p_2 - \beta_y p_0)\\ p_3 \end{pmatrix}\]

where the kinematic factor is given as $\gamma = 1/\sqrt{1-\beta_y^2}$)

Example

julia> using QEDcore

julia> using Random

julia> RNG = MersenneTwister(1234)
MersenneTwister(1234)

julia> beta_y = BetaY(0.5)
BetaY{Float64}(0.5)

julia> boost = Boost(beta_y)
Boost{BetaY{Float64}}(BetaY{Float64}(0.5))

julia> p = SFourMomentum(4,3,2,1)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> p_prime = boost(p)
4-element SFourMomentum with indices SOneTo(4):
 3.4641016151377553
 3.0
 0.0
 1.0

julia> @assert isapprox(p*p,p_prime*p_prime) # Invariant mass is preserved

External link

Lorentz Boost on Wikipedia

Kinematics in PDG review

QEDcore.BetaZ — Type

BetaZ(beta::T) where {T<:Real}

Represents the beta parameter associated with a Lorentz boost along the z-axis, commonly denoted as $\beta_z$.

The transformation for a boost along the z-axis is:

\[\begin{pmatrix} p_0\\ p_1\\ p_2\\ p_3 \end{pmatrix} \mapsto \begin{pmatrix} \gamma (p_0 - \beta_z p_3)\\ p_1\\ p_2\\ \gamma (p_3 - \beta_z p_0)\\ \end{pmatrix}\]

where the kinematic factor is given as $\gamma = 1/\sqrt{1-\beta_z^2}$)

Example

julia> using QEDcore

julia> using Random

julia> RNG = MersenneTwister(1234)
MersenneTwister(1234)

julia> beta_z = BetaZ(0.5)
BetaZ{Float64}(0.5)

julia> boost = Boost(beta_z)
Boost{BetaZ{Float64}}(BetaZ{Float64}(0.5))

julia> p = SFourMomentum(4,3,2,1)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> p_prime = boost(p)
4-element SFourMomentum with indices SOneTo(4):
  4.041451884327381
  3.0
  2.0
 -1.1547005383792517

julia> @assert isapprox(p*p,p_prime*p_prime) # Invariant mass is preserved

External link

Lorentz Boost on Wikipedia

Kinematics in PDG review

QEDcore.BiSpinor — Type

struct BiSpinor <: AbstractDiracVector{ComplexF64}

Concrete type to model a Dirac four-spinor with complex-valued components. These are the elements of an actual spinor space.

QEDcore.Boost — Type

Boost{V<:AbstractBoostParameter} <: AbstractLorentzBoost

A concrete type representing a Lorentz boost transformation, parameterized by a boost parameter V. The boost parameter can be either axis-specific or vector-like, depending on the subtype of AbstractBoostParameter used. The Boost type is used to perform Lorentz boosts on four-vectors (such as four-momentum or four-position) between different inertial frames in special relativity.

Fields

param::V: A boost parameter of type V, which is a subtype of AbstractBoostParameter. This parameter defines the velocity (as a fraction of the speed of light, $\beta$) and the direction of the boost (e.g., along a single axis or in multiple directions).

Overview

A Lorentz boost is a transformation that adjusts the time and spatial components of a four-vector based on the relative velocity between two reference frames. The Boost struct provides a general and flexible implementation of such a boost, where the type of the boost parameter determines the direction and magnitude of the boost.

Depending on the boost parameter V, the boost can be:

Axis-specific: When V is an axis-specific boost parameter (e.g., BetaX), the boost will be along that axis.

Vector-like: When V is a vector of boost parameters (e.g., BetaVector), the boost will have components in multiple spatial directions.

Example

To create a Lorentz boost along the x-axis using the BetaX boost parameter:

julia> using QEDcore

julia> beta_x = BetaX(0.5)
BetaX{Float64}(0.5)

julia> boost_x = Boost(beta_x)
Boost{BetaX{Float64}}(BetaX{Float64}(0.5))

To perform a Lorentz boost using the boost_x object, you can apply it to a four-vector, such as four-momentum:

julia> p = SFourMomentum(4, 3, 2, 1)
4-element SFourMomentum with indices SOneTo(4):
 4.0
 3.0
 2.0
 1.0

julia> p_prime = boost_x(p)  # Perform the boost
4-element SFourMomentum with indices SOneTo(4):
 2.886751345948129
 1.1547005383792517
 2.0
 1.0

julia> @assert isapprox(p*p, p_prime*p_prime)  # The invariant mass is preserved

Notes

The Boost type provides a unified and flexible interface for applying Lorentz boosts, with the boost parameter V determining the specific form of the transformation. Lorentz boosts preserve the spacetime interval, meaning that applying the boost to a four-vector will not change the invariant quantity.

See Also

AbstractBoostParameter: Base type for specific kinds of boost parameters.

BetaX: Boost parameter for the x-axis.

BetaY: Boost parameter for the y-axis.

BetaZ: Boost parameter for the z-axis.

BetaVector: Vector of boost parameters for boosts in multiple spatial directions.

QEDcore.Boson — Type

Abstract base type for bosons as distinct from its anti-particle counterpart AntiBoson.

particle interface

All subtypes of Boson have

is_boson(::Boson) = true
is_particle(::Boson) = true
is_anti_particle(::Boson) = false

QEDcore.BosonLike — Type

Abstract base types for particle species that act like bosons in the sense of particle statistics.

particle interface

Every concrete subtype of BosonLike has is_boson(::BosonLike) = true.

QEDcore.CenterOfMomentumFrame — Type

CenterOfMomentumFrame <: AbstractFrameOfReference

TBW

QEDcore.DiracMatrix — Type

struct DiracMatrix <: AbstractDiracMatrix{ComplexF64}

Concrete type to model Dirac matrices, i.e. matrix representations of linear mappings between two spinor spaces.

QEDcore.Electron — Type

Concrete type for electrons as a particle species. Mostly used for dispatch.

julia> using QEDcore

julia> Electron()
electron

particle interface

Besides being a subtype of Fermion, objects of type Electron have

mass(::Electron) = 1.0
charge(::Electron) = -1.0

QEDcore.ElectronRestFrame — Type

ElectronRestFrame <: AbstractFrameOfReference

TBW

QEDcore.Fermion — Type

Abstract base type for fermions as distinct from AntiFermions.

particle interface

All subtypes of Fermion have

is_fermion(::Fermion) = true
is_particle(::Fermion) = true
is_anti_particle(::Fermion) = false

QEDcore.FermionLike — Type

Abstract base types for particle species that act like fermions in the sense of particle statistics.

particle interface

Every concrete subtype of FermionLike has is_fermion(::FermionLike) = true.

QEDcore.InPhaseSpacePoint — Type

InPhaseSpacePoint

A partial type specialization on PhaseSpacePoint which can be used for dispatch in functions requiring only the in channel of the phase space to exist, for example implementations of _incident_flux. No restrictions are imposed on the out-channel, which may or may not exist.

See also: OutPhaseSpacePoint

QEDcore.InPhaseSpacePoint — Method

InPhaseSpacePoint(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    ps_def::AbstractPhasespaceDefinition,
    in_momenta::NTuple{N,AbstractFourMomentum},
)

Construct a PhaseSpacePoint with only input particles from given momenta. The result will be <: InPhaseSpacePoint but not <: OutPhaseSpacePoint.

QEDcore.InPhaseSpacePoint — Method

InPhaseSpacePoint(
    proc::AbstractProcessDefinition, 
    model::AbstractModelDefinition, 
    ps_def::AbstractPhasespaceDefinition, 
    in_ps::Tuple{ParticleStateful},
)

Construct a [`PhaseSpacePoint`](@ref) with only input particles from [`ParticleStateful`](@ref)s. The result will be `<: InPhaseSpacePoint` but **not** `<: OutPhaseSpacePoint`.

QEDcore.InPhaseSpacePoint — Method

InPhaseSpacePoint(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    ps_def::AbstractPhasespaceDefinition,
    in_coords::NTuple{N,Real},
)

Construct a PhaseSpacePoint from given coordinates by using the _generate_momenta interface. The result will be <: InPhaseSpacePoint but not <: OutPhaseSpacePoint.

Note

A similar function for OutPhaseSpacePoint does not exist from coordinates, only a full PhaseSpacePoint.

QEDcore.MFourMomentum — Type

mutable struct MFourMomentum <: AbstractFourMomentum

Builds a mutable LorentzVector with real components used to statically model the four-momentum of a particle or field.

Fields

E::Float64: energy component

px::Float64: x component

py::Float64: y component

pz::Float64: z component

QEDcore.MFourMomentum — Method

MFourMomentum(t, x, y, z)

The interface transforms each number-like input to float64:

MFourMomentum(
    t::Union{Integer, Irrational, Rational},
    x::Union{Integer, Irrational, Rational},
    y::Union{Integer, Irrational, Rational},
    z::Union{Integer, Irrational, Rational}
) -> MFourMomentum

QEDcore.MLorentzVector — Type

mutable struct MLorentzVector{T} <: AbstractLorentzVector{T}

Concrete implementation of a generic mutable Lorentz vector. Each manipulation of an concrete implementation which is not self-contained (i.e. produces the same Lorentz vector type) will result in this type.

Fields

t::Any: t component

x::Any: x component

y::Any: y component

z::Any: z component

QEDcore.MajoranaBoson — Type

Abstract base type for majorana-bosons, i.e. bosons which are their own anti-particles.

particle interface

All subtypes of MajoranaBoson have

is_boson(::MajoranaBoson) = true
is_particle(::MajoranaBoson) = true
is_anti_particle(::MajoranaBoson) = true

QEDcore.MajoranaFermion — Type

Abstract base type for majorana-fermions, i.e. fermions which are their own anti-particles.

particle interface

All subtypes of MajoranaFermion have

is_fermion(::MajoranaFermion) = true
is_particle(::MajoranaFermion) = true
is_anti_particle(::MajoranaFermion) = true

QEDcore.OutPhaseSpacePoint — Type

OutPhaseSpacePoint

A partial type specialization on PhaseSpacePoint which can be used for dispatch in functions requiring only the out channel of the phase space to exist. No restrictions are imposed on the in-channel, which may or may not exist.

See also: InPhaseSpacePoint

QEDcore.OutPhaseSpacePoint — Method

OutPhaseSpacePoint(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    ps_def::AbstractPhasespaceDefinition,
    out_momenta::NTuple{N,AbstractFourMomentum},
)

Construct a PhaseSpacePoint with only output particles from given momenta. The result will be <: OutPhaseSpacePoint but not <: InPhaseSpacePoint.

QEDcore.OutPhaseSpacePoint — Method

OutPhaseSpacePoint(
    proc::AbstractProcessDefinition, 
    model::AbstractModelDefinition, 
    ps_def::AbstractPhasespaceDefinition, 
    out_ps::Tuple{ParticleStateful},
)

Construct a PhaseSpacePoint with only output particles from ParticleStatefuls. The result will be <: OutPhaseSpacePoint but not <: InPhaseSpacePoint.

QEDcore.ParticleStateful — Type

ParticleStateful <: AbstractParticle

Representation of a particle with a state. It has four fields:

dir::ParticleDirection: The direction of the particle, Incoming() or Outgoing().

species::AbstractParticleType: The species of the particle, Electron(), Positron() etc.

mom::AbstractFourMomentum: The momentum of the particle.

Overloads for is_fermion, is_boson, is_particle, is_anti_particle, is_incoming, is_outgoing, mass, and charge are provided, delegating the call to the correct field and thus implementing the AbstractParticle interface.

julia> using QEDcore

julia> ParticleStateful(Incoming(), Electron(), SFourMomentum(1, 0, 0, 0))
ParticleStateful: incoming electron
    momentum: [1.0, 0.0, 0.0, 0.0]

julia> ParticleStateful(Outgoing(), Photon(), SFourMomentum(1, 0, 0, 0))
ParticleStateful: outgoing photon
    momentum: [1.0, 0.0, 0.0, 0.0]

QEDcore.ParticleStateful — Method

ParticleStateful{DIR, SPECIES}(mom::AbstractFourMomentum)
ParticleStateful{DIR, SPECIES, EL}(mom::EL)

Construct a ParticleStateful from the given momentum on a fully or partially specified type.

QEDcore.PhaseSpacePoint — Type

PhaseSpacePoint

Representation of a point in the phase space of a process. Contains the process (AbstractProcessDefinition), the model (AbstractModelDefinition), the phase space definition (AbstractPhasespaceDefinition), and stateful incoming and outgoing particles (ParticleStateful).

The legality of the combination of the given process and the incoming and outgoing particles is checked on construction. If the numbers of particles mismatch, the types of particles mismatch (note that order is important), or incoming particles have an Outgoing direction, an error is thrown.

julia> using QEDcore; using QEDprocesses

julia> PhaseSpacePoint(
            Compton(),
            PerturbativeQED(),
            PhasespaceDefinition(SphericalCoordinateSystem(), ElectronRestFrame()),
            (
                ParticleStateful(Incoming(), Electron(), SFourMomentum(1, 0, 0, 0)),
                ParticleStateful(Incoming(), Photon(), SFourMomentum(1, 0, 0, 0))
            ),
            (
                ParticleStateful(Outgoing(), Electron(), SFourMomentum(1, 0, 0, 0)),
                ParticleStateful(Outgoing(), Photon(), SFourMomentum(1, 0, 0, 0))
            )
        )
PhaseSpacePoint:
    process: one-photon Compton scattering
    model: perturbative QED
    phasespace definition: spherical coordinates in electron rest frame
    incoming particles:
     -> incoming electron: [1.0, 0.0, 0.0, 0.0]
     -> incoming photon: [1.0, 0.0, 0.0, 0.0]
    outgoing particles:
     -> outgoing electron: [1.0, 0.0, 0.0, 0.0]
     -> outgoing photon: [1.0, 0.0, 0.0, 0.0]

Note

PhaseSpacePoints can be constructed with only one of their in- or out-channel set. For this, see the special constructors InPhaseSpacePoint and OutPhaseSpacePoint. The InPhaseSpacePoint and OutPhaseSpacePoint type definitions can be used to dispatch on such PhaseSpacePoints. Note that a full PhaseSpacePoint containing both its in- and out-channel matches both, .i.e. psp isa InPhaseSpacePoint and psp isa OutPhaseSpacePoint both evaluate to true if psp contains both channels. A completely empty PhaseSpacePoint is not allowed.

QEDcore.PhaseSpacePoint — Method

PhaseSpacePoint(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    ps_def::AbstractPhasespaceDefinition,
    in_momenta::NTuple{N,AbstractFourMomentum},
    out_momenta::NTuple{M,AbstractFourMomentum},
)

Construct the phase space point from given momenta of incoming and outgoing particles regarding a given process.

QEDcore.PhaseSpacePoint — Method

PhaseSpacePoint(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    ps_def::AbstractPhasespaceDefinition,
    in_coords::NTuple{N,Real},
    out_coords::NTuple{M,Real},
)

Construct a PhaseSpacePoint from given coordinates by using the _generate_momenta interface.

QEDcore.PhasespaceDefinition — Type

PhasespaceDefinition(coord_sys::AbstractCoordinateSystem, frame::AbstractFrameOfReference)

Convenient type to dispatch on coordiante systems and frames of reference. Combines a AbstractCoordinateSystem with a AbstractFrameOfReference.

QEDcore.Photon — Type

Concrete type for the photons as a particle species. Mostly used for dispatch.

julia> using QEDcore

julia> Photon()
photon

particle interface

Besides being a subtype of MajoranaBoson, Photon has

mass(::Photon) = 0.0
charge(::Photon) = 0.0

QEDcore.Positron — Type

Concrete type for positrons as a particle species. Mostly used for dispatch.

julia> using QEDcore

julia> Positron()
positron

particle interface

Besides being a subtype of AntiFermion, objects of type Positron have

mass(::Positron) = 1.0
charge(::Positron) = 1.0

QEDcore.SFourMomentum — Type

struct SFourMomentum <: AbstractFourMomentum

Builds a static LorentzVectorLike with real components used to statically model the four-momentum of a particle or field.

Fields

E::Float64: energy component

px::Float64: x component

py::Float64: y component

pz::Float64: z component

QEDcore.SFourMomentum — Method

SFourMomentum(t, x, y, z)

The interface transforms each number-like input to float64:

SFourMomentum(
    t::Union{Integer, Irrational, Rational},
    x::Union{Integer, Irrational, Rational},
    y::Union{Integer, Irrational, Rational},
    z::Union{Integer, Irrational, Rational}
) -> SFourMomentum

QEDcore.SLorentzVector — Type

struct SLorentzVector{T} <: AbstractLorentzVector{T}

Concrete implementation of a generic static Lorentz vector. Each manipulation of an concrete implementation which is not self-contained (i.e. produces the same Lorentz vector type) will result in this type.

Fields

t::Any: t component

x::Any: x component

y::Any: y component

z::Any: z component

QEDcore.SphericalCoordinateSystem — Type

SphericalCoordinateSystem <: AbstractCoordinateSystem

TBW

Base.getindex — Method

Base.getindex(psp::PhaseSpacePoint, dir::Incoming, n::Int)

Overload for the array indexing operator []. Returns the nth incoming particle in this phase space point.

Base.getindex — Method

Base.getindex(psp::PhaseSpacePoint, dir::Outgoing, n::Int)

Overload for the array indexing operator []. Returns the nth outgoing particle in this phase space point.

QEDcore._momentum_type — Method

_momentum_type(psp::PhaseSpacePoint)
_momentum_type(type::Type{PhaseSpacePoint})

Returns the element type of the PhaseSpacePoint object or type, e.g. SFourMomentum.

julia> using QEDcore; using QEDprocesses

julia> psp = PhaseSpacePoint(Compton(), PerturbativeQED(), PhasespaceDefinition(SphericalCoordinateSystem(), ElectronRestFrame()), Tuple(rand(SFourMomentum) for _ in 1:2), Tuple(rand(SFourMomentum) for _ in 1:2));

julia> QEDcore._momentum_type(psp)
SFourMomentum

julia> QEDcore._momentum_type(typeof(psp))
SFourMomentum

QEDcore._mul — Method

_mul(aBS::AdjointBiSpinor, BS::BiSpinor) -> ComplexF64

Tensor product of an adjoint with a standard bi-spinor resulting in a scalar.

Multiplication operator

This also overloads the * operator for this types.

QEDcore._mul — Method

_mul(
    aBS::AdjointBiSpinor,
    DM::DiracMatrix,
    BS::BiSpinor
) -> ComplexF64

Tensor product of Dirac matrix sandwiched between an adjoint and a standard bi-spinor resulting in a scalar.

QEDcore._mul — Method

_mul(
    aBS::AdjointBiSpinor,
    DM::DiracMatrix
) -> AdjointBiSpinor

Tensor product of an adjoint bi-spinor with a Dirac matrix resulting in another adjoint bi-spinor.

Multiplication operator

This also overloads the * operator for this types.

QEDcore._mul — Method

_mul(BS::BiSpinor, aBS::AdjointBiSpinor) -> DiracMatrix

Tensor product of a standard with an adjoint bi-spinor resulting in a Dirac matrix.

Multiplication operator

This also overloads the * operator for this types.

QEDcore._mul — Method

_mul(DM::DiracMatrix, BS::BiSpinor) -> BiSpinor

Tensor product of an Dirac matrix with a standard bi-spinor resulting in another standard bi-spinor.

Multiplication operator

This also overloads the * operator for this types.

QEDcore._mul — Method

_mul(DM1::DiracMatrix, DM2::DiracMatrix) -> DiracMatrix

Tensor product two Dirac matrices resulting in another Dirac matrix.

Multiplication operator

This also overloads the * operator for this types.

QEDcore._transform — Function

_transform(trafo::AbstractCoordinateTransformation,p::AbstractFourMomentum)

Interface function for the application of the transformation to the four-momentum p. Must return a four-momentum of the same type as p.