Phase Space Layout

Phase Space Layout Interface

QEDcore.AbstractPhaseSpaceLayout — Type

AbstractPhaseSpaceLayout

The AbstractPhaseSpaceLayout is an abstract type that represents the general concept of a phase space layout in a scattering process.

Interface Functions to Implement:

phase_space_dimension(proc, model, layout::AbstractPhaseSpaceLayout): Defines the number of independent phase space coordinates needed to build the momenta.

QEDcore.AbstractInPhaseSpaceLayout — Type

AbstractInPhaseSpaceLayout <: AbstractPhaseSpaceLayout

The AbstractInPhaseSpaceLayout represents the phase space layout for the incoming particles in a scattering process. It defines the way in which the momenta of incoming particles are constructed from phase space coordinates.

Interface Functions to Implement:

phase_space_dimension(proc, model, layout::AbstractInPhaseSpaceLayout): Defines the number of independent phase space coordinates for the incoming particles.

build_momenta(proc, model, in_psl::AbstractInPhaseSpaceLayout, in_coords::Tuple): Constructs the momenta for the incoming particles using the phase space coordinates.

QEDcore.AbstractOutPhaseSpaceLayout — Type

AbstractOutPhaseSpaceLayout{InPSL<:AbstractInPhaseSpaceLayout} <: AbstractPhaseSpaceLayout

The AbstractOutPhaseSpaceLayout represents the phase space layout for the outgoing particles in a scattering process. It typically depends on the phase space layout of the incoming particles, and it specifies how the momenta of the outgoing particles are constructed from the respective coordinates.

The generic parameter InPSL links the outgoing phase space layout to the incoming layout, allowing consistency between the two configurations in the process.

Interface Functions to Implement:

phase_space_dimension(proc, model, layout::AbstractOutPhaseSpaceLayout): Defines the number of independent phase space coordinates for the outgoing particles.

in_phase_space_layout(out_psl::AbstractOutPhaseSpaceLayout): Provides the associated incoming phase space layout to ensure consistency between incoming and outgoing configurations.

build_momenta(proc, model, Ptot::AbstractFourMomentum, out_psl::AbstractOutPhaseSpaceLayout, out_coords::Tuple): Constructs the momenta for the outgoing particles, ensuring they comply with energy and momentum conservation based on the total incoming four-momentum.

QEDcore._build_momenta — Function

_build_momenta(proc, model, in_psl::AbstractInPhaseSpaceLayout, in_coords::Tuple)
_build_momenta(proc, model, Ptot::AbstractFourMomentum, out_psl::AbstractOutPhaseSpaceLayout, out_coords::Tuple)

These functions need to be implemented as part of the phase space layout interface for both incoming and outgoing particle momenta construction. They serve as internal, low-level interfaces for constructing the four-momenta of particles during a scattering process, and are typically wrapped by the user-facing build_momenta function.

Incoming Phase Space Layout

The first function, _build_momenta(proc, model, in_psl, in_coords), constructs the four-momenta for the incoming particles based on the specified phase space coordinates (in_coords).

Arguments:

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

in_psl: The incoming phase space layout, subtype of AbstractInPhaseSpaceLayout, that defines how to map the coordinates to momenta.

in_coords: A tuple of phase space coordinates that parametrize the momenta of the incoming particles.

Returns:

A collection of four-momenta representing the incoming particles. For performance reasons, it is recommended to return a Tuple of four-momenta.

Outgoing Phase Space Layout

The second function, _build_momenta(proc, model, Ptot, out_psl, out_coords), constructs the four-momenta for the outgoing particles. It uses the incoming four-momenta (in_moms) from the incoming state and applies the phase space coordinates (out_coords) to compute the outgoing momenta, ensuring they adhere to energy and momentum conservation laws.

Arguments:

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

in_moms: The incoming four-momenta, which is used to compute the momenta of the outgoing particles.

out_psl: The outgoing phase space layout, subtype of AbstractOutPhaseSpaceLayout, that maps the coordinates to momenta.

out_coords: A tuple of phase space coordinates that parametrize the outgoing particle momenta.

Returns:

A collection of four-momenta representing the outgoing particles. For performance reasons, it is recommended to return a Tuple of four-momenta.

Notes

Both versions of _build_momenta handle the construction of particle momenta during different phases of the scattering process:

The incoming version constructs momenta based on phase space coordinates alone.

The outgoing version constructs momenta based on both phase space coordinates and the total four-momentum, ensuring conservation laws are respected.

QEDcore.phase_space_dimension — Function

phase_space_dimension(proc, model, layout::AbstractPhaseSpaceLayout) -> Int

This function needs to be implemented for the phase-space layout interface. Return the dimensionality of the phase space, i.e. the numebr of coordinates, for a given process and model within the specified layout.

The phase space dimension is a crucial quantity that determines how many independent coordinates are required to describe the system of particles in the scattering process. It depends on the number of particles involved and the specific interaction model in use.

Arguments

proc: The scattering process definition, a subtype of AbstractProcessDefinition.

model: The physics model, a subtype of AbstractModelDefinition.

layout: A specific phase space layout, either AbstractInPhaseSpaceLayout or AbstractOutPhaseSpaceLayout.

Returns

The integer representing the number of independent phase space coordinates.

QEDcore.in_phase_space_layout — Function

in_phase_space_layout(out_psl::AbstractOutPhaseSpaceLayout) -> AbstractInPhaseSpaceLayout

This function needs to be implemented for the out-phase-space layout interface. Given an outgoing phase space layout (out_psl), this function returns the associated incoming phase space layout.

This is useful for ensuring consistency between the incoming and outgoing particle momenta when calculating or sampling phase space points in scattering processes.

Arguments

out_psl: The outgoing phase space layout, a subtype of AbstractOutPhaseSpaceLayout.

Returns

The associated incoming phase space layout, a subtype of AbstractInPhaseSpaceLayout.

Building Momenta

QEDcore.build_momenta — Function

build_momenta(proc, model, in_psl::AbstractInPhaseSpaceLayout, in_coords::Tuple)

Construct the momenta of the incoming particles using the provided phase space coordinates.

This is the user-facing function that calls _build_momenta internally and validates the number of coordinates against the phase space dimensionality.

Arguments

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

in_psl: The incoming phase space layout, subtype of AbstractInPhaseSpaceLayout.

in_coords: A tuple of phase space coordinates that parametrize the incoming particle momenta.

Returns

A collection of four-momenta representing the incoming particles. Because of performance reasons, it is recommened to return a Tuple of four-momenta.

build_momenta(proc::AbstractProcessDefinition, model::AbstractModelDefinition, in_psl::AbstractInPhaseSpaceLayout, in_coords::Real)

A scalar version of build_momenta for incoming phase space layouts (in_psl), where the phase space coordinates are provided as a single scalar instead of a tuple.

Arguments:

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

in_psl: The incoming phase space layout, subtype of AbstractInPhaseSpaceLayout.

in_coords::Real: A single scalar representing the phase space coordinate for the incoming particles.

Returns:

A collection of four-momenta representing the incoming particles. Because of performance reasons, it is recommened to return a Tuple of four-momenta.

Notes:

This function is a convenience wrapper around build_momenta, automatically converting the scalar in_coords into a 1-tuple. It is useful when the incoming phase space only requires a single coordinate to define the particle momenta.

build_momenta(proc, model, Ptot::AbstractFourMomentum, out_psl::AbstractOutPhaseSpaceLayout, out_coords::Tuple)

Construct the momenta of the outgoing particles using the provided phase space coordinates (out_coords) and total incoming momentum (Ptot).

This function ensures that the outgoing momenta satisfy energy and momentum conservation, consistent with the physics model in use.

Arguments

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

in_moms: The incoming four-momenta, used to compute the outgoing momenta.

out_psl: The outgoing phase space layout, subtype of `AbstractOutPhaseSpaceLayout.

out_coords: A tuple of phase space coordinates that parametrize the outgoing particle momenta.

Returns

A collection of four-momenta representing the incoming particles. Because of performance reasons, it is recommened to return a Tuple of four-momenta.

Coordinate Maps

QEDcore.CoordinateMap — Type

CoordinateMap{P,M,PSL}(proc::P, model::M, psl::PSL)

A CoordinateMap represents a transformation that maps phase space coordinates to particle momenta for a specific scattering process. This type encapsulates the scattering process definition (proc), the physics model (model), and the phase space layout (psl). The CoordinateMap is callable and supports the conversion of both incoming and outgoing phase space coordinates into four-momenta.

Fields

proc: The scattering process definition, subtype of AbstractProcessDefinition.

model: The physics model, subtype of AbstractModelDefinition.

psl: The phase space layout, either incoming or outgoing, that maps coordinates to momenta.

Usage

Incoming Coordinates

The CoordinateMap can be called with a tuple of incoming phase space coordinates (in_coords), which are used to compute the corresponding incoming particle four-momenta. This is done by calling the phase space construction function build_momenta using the proc, model, and psl provided in the CoordinateMap.

# Create a coordinate map for in phase space
in_coord_map = CoordinateMap(proc, model, in_psl)

# call on in-coordinates to build the momenta
in_coord_map(in_coords)

Arguments:

in_coords: A tuple of phase space coordinates for the incoming particles.

Returns: A tuple of four-momenta corresponding to the incoming particles.

Incoming and Outgoing Coordinates

The CoordinateMap can also be called with both incoming (in_coords) and outgoing (out_coords) phase space coordinates. The incoming momenta are computed first, and then the total momentum (Ptot) is calculated by summing these momenta. This total momentum is then used to compute the outgoing particle four-momenta using the provided outgoing phase space layout.

# Create a coordinate map for out phase space
coord_map = CoordinateMap(proc, model, out_psl)

# call on in- and out-coordinates to build the momenta
coord_map(in_coords, out_coords)

Arguments:

in_coords: A tuple of phase space coordinates for the incoming particles.

out_coords: A tuple of phase space coordinates for the outgoing particles.

Returns:

A tuple where the first element is the incoming particle momenta and the second element is the outgoing particle momenta, both represented as tuples of four-momenta.

Notes

The CoordinateMap provides a flexible mechanism for transforming phase space coordinates into physically meaningful four-momenta, ensuring consistency with the scattering process, physics model, and phase space layout.

The type is designed to handle both incoming and outgoing momenta, ensuring proper energy and momentum conservation for the process.

QEDcore.CoordinateMapCached — Type

CoordinateMapCached{P,M,PSL,TM}(proc::P, model::M, psl::PSL, in_moms::TM)

A CoordinateMapCached represents a precomputed transformation for phase space coordinates, where the momenta of the incoming particles are cached. This can improve performance in cases where the incoming momenta remain constant and only the outgoing momenta need to be computed based on new coordinates.

The cached map encapsulates the scattering process (proc), the physics model (model), the phase space layout (psl), and the precomputed incoming momenta (in_moms).

Fields

proc: The scattering process definition, which defines the incoming and outgoing particles.

model: The physics model, which governs the interaction type and momentum distributions.

psl: The phase space layout, either incoming or outgoing, that maps coordinates to momenta.

in_moms: A collection of precomputed four-momenta for the incoming particles, usually a Tuple.

Usage

Cached Incoming Coordinates

When a CoordinateMapCached build with psl::AbstractInPhaseSpaceLayout is called without any arguments, it returns the precomputed incoming momenta (in_moms) directly. This provides an efficient way to access the incoming particle momenta that have already been calculated and stored in the map.

# Create a coordinate map for in phase space
in_coord_map = CoordinateMapCached(proc, model, in_psl, in_moms)

# call to return the in-momenta
in_coord_map()

Returns: The cached tuple of four-momenta for the incoming particles.

Outgoing Coordinates

The CoordinateMapCached can be called with outgoing phase space coordinates (out_coords). The cached incoming momenta (in_moms) are used to compute the total momentum (Ptot), which is then passed along with the outgoing coordinates to compute the momenta of the outgoing particles.

# Create a coordinate map for out phase space
out_coord_map = CoordinateMapCached(proc, model, out_psl, in_moms)

# call on out coordinates to return the out-momenta
out_coord_map(out_coords)

Arguments:

out_coords: A tuple of phase space coordinates for the outgoing particles.

Returns:

A tuple of four-momenta for the outgoing particles, consistent with the total momentum derived from the cached incoming momenta.

Notes

Caching: The CoordinateMapCached is useful when the incoming momenta are fixed or do not change frequently, as it avoids recomputation by storing the incoming momenta in the cache.

Efficiency: This caching mechanism can significantly enhance performance when repeatedly evaluating the scattering process with fixed incoming particles but varying outgoing configurations.

The type is designed to handle both incoming and outgoing momenta, ensuring proper energy and momentum conservation for the process.

Implementations

Two-body Phase Space Layout

QEDcore.AbstractTwoBodyInPhaseSpaceLayout — Type

AbstractTwoBodyInPhaseSpaceLayout <: AbstractInPhaseSpaceLayout

An abstract type representing an incoming phase space layout specifically designed for two-body systems in high-energy physics. This type is a specialized subtype of AbstractInPhaseSpaceLayout, focusing on scenarios where two particles are incoming and participate in a scattering or decay process.

Concrete subtypes of AbstractTwoBodyInPhaseSpaceLayout define the parameterization of the momenta of the two incoming particles, such as by Energy, Rapidity, or other kinematic coordinates. These layouts allow for consistent calculation of the system's phase-space properties under conservation laws, facilitating method dispatch in functions like _build_momenta.

This type serves as a foundation for implementing concrete layouts, ensuring that two-body processes can be modeled flexibly and accurately within phase space calculations.

See Also

AbstractInPhaseSpaceLayout: A broader type for general incoming phase space layouts.

AbstractTwoBodyRestSystem: A subtype representing two-body systems where one particle is at rest.

QEDcore.AbstractTwoBodyRestSystem — Type

AbstractTwoBodyRestSystem <: AbstractTwoBodyInPhaseSpaceLayout

An abstract type representing the phase space layout of a two-body system in which one particle is at rest in the system's frame.

Concrete subtypes of AbstractTwoBodyRestSystem typically specify the coordinate system used to parameterize the momentum of the non-resting particle, such as Energy, SpatialMagnitude, CenterOfMomentumEnergy, or Rapidity. The choice of coordinate system impacts how the momenta of the incoming particles are constructed.

This type serves as a base for defining custom layouts that rely on rest-frame assumptions for two-body scattering or decay processes, and it facilitates method dispatch in momentum construction functions like _build_momenta.

See Also

TwoBodyRestSystem: A concrete implementation of AbstractTwoBodyRestSystem that supports various univariate coordinates for describing the non-resting particle.

QEDcore.TwoBodyRestSystem — Type

TwoBodyRestSystem{RESTIDX, COORD<:AbstractUnivariateCoordinate}

Represents a two-body scattering system in the rest frame of one of the particles, where one particle (identified by RESTIDX) is at rest and the other particle's momentum is described by a coordinate system COORD. The system uses various univariate coordinates, such as energy, rapidity, or center-of-momentum energy, to parameterize the momentum of the moving particle.

This type allows for easy construction of the incoming particle momenta using the phase space layout for two-body systems, commonly used in high-energy physics processes. The system supports different coordinate systems for defining the moving particle's momenta, including energy-based and rapidity-based layouts.

Fields

coord::COORD: The coordinate type (COORD) used to describe the non-resting particle's momenta.

Supported Coordinates

Energy: Defines the energy of the moving particle.

SpatialMagnitude: Defines the spatial momentum magnitude of the moving particle.

CMSEnergy: Defines the total center-of-mass energy of the system.

Rapidity: Defines the rapidity of the moving particle.

Example Usages

The following examples show how to use different coordinate systems with TwoBodyRestSystem in conjunction with the _build_momenta interface.

Example 1: Using Energy Coordinate

psl = TwoBodyRestSystem(Energy(2))  # Particle 1 at rest, particle 2 described by its energy
in_coords = (100.0,)  # Energy of particle 2
momenta = build_momenta(proc, model, psl, in_coords)

This constructs the momenta assuming particle 1 is at rest, and particle 2 has an energy of 100 GeV.

Example 2: Using SpatialMagnitude Coordinate

psl = TwoBodyRestSystem(SpatialMagnitude(2))  # Particle 1 at rest, particle 2 described by its spatial magnitude
in_coords = (50.0,)  # Spatial momentum magnitude of particle 2
momenta = build_momenta(proc, model, psl, in_coords)

In this case, the moving particle's spatial momentum magnitude is given, and its energy is calculated based on its mass.

Example 3: Using CMSEnergy Coordinate

psl = TwoBodyRestSystem(1,CMSEnergy())  # Particle 1 at rest, particle 2 described by the center-of-mass energy
in_coords = (200.0,)  # Total center-of-mass energy of the system
momenta = build_momenta(proc, model, psl, in_coords)

Here, the center-of-mass energy of the system is provided, and the momenta of both particles are calculated while conserving energy and momentum.

Example 4: Using Rapidity Coordinate

psl = TwoBodyRestSystem(Rapidity(2))  # Particle 1 at rest, particle 2 described by its rapidity
in_coords = (1.0,)  # Rapidity of particle 2
momenta = _build_momenta(proc, model, psl, in_coords)
@assert isapprox(getMass(sum(momenta)),50.0)

In this case, the rapidity of particle 2 is given, and its energy and momentum are computed based on this parameter.

Notes

RESTIDX is the index of the particle that is at rest in the system.

COORD specifies the coordinate type of the non-resting particle and must be a subtype of AbstractUnivariateCoordinate.

The coordinate system should be compatible with the given particles and their masses, otherwise, an error may occur during momentum construction.

Throws

ArgumentError if invalid coordinates are used (e.g., unsupported coordinate types or incompatible indices).